+++ modules/pytket-qiskit/tests/backend_test.py
@@ -1,4 +1,4 @@
-# Copyright 2019-2021 Cambridge Quantum Computing
+# Copyright 2019-2022 Cambridge Quantum Computing
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -41,7 +41,6 @@ from pytket.extensions.qiskit import (
     AerUnitaryBackend,
     IBMQEmulatorBackend,
 )
-from pytket.extensions.qiskit.backends.ibm import _rebase_pass
 from pytket.extensions.qiskit import qiskit_to_tk, process_characterisation
 from pytket.utils.expectations import (
     get_pauli_expectation_value,
@@ -312,17 +311,20 @@ def test_process_characterisation_complete_noise_model() -> None:
 
     gqe2 = cast(Dict, char["GenericTwoQubitQErrors"])
     gqe1 = cast(Dict, char["GenericOneQubitQErrors"])
-    assert gqe2[(0, 1)][0][1][0] == 0.0375
-    assert gqe2[(0, 1)][0][1][15] == 0.4375
-    assert gqe1[0][0][1][0] == 0.125
-    assert gqe1[0][0][1][3] == 0.625
+
+    assert round(gqe2[(0, 1)][0][1][15], 5) == 0.0375
+    assert round(gqe2[(0, 1)][0][1][0], 5) == 0.4375
+    assert gqe1[0][0][1][3] == 0.125
+    assert gqe1[0][0][1][0] == 0.625
     assert gqe1[0][1][1][0] == 0.35
     assert gqe1[0][1][1][1] == 0.65
     assert gqe1[0][2][1][0] == 0.35
     assert gqe1[0][2][1][1] == 0.65
     assert node_errors[arch.nodes[0]][OpType.U3] == 0.375
-    assert link_errors[(arch.nodes[0], arch.nodes[1])][OpType.CX] == 0.5625
-    assert link_errors[(arch.nodes[1], arch.nodes[0])][OpType.CX] == 0.80859375
+    assert round(link_errors[(arch.nodes[0], arch.nodes[1])][OpType.CX], 4) == 0.5625
+    assert (
+        round(link_errors[(arch.nodes[1], arch.nodes[0])][OpType.CX], 8) == 0.80859375
+    )
     readout_errors = cast(Dict, back.backend_info.all_readout_errors)
     assert readout_errors[arch.nodes[0]] == [
         [0.8, 0.2],
@@ -983,7 +985,7 @@ def test_symbolic_rebase() -> None:
     pytket_circ = qiskit_to_tk(circ)
 
     # rebase pass could not handle symbolic parameters originally and would fail here:
-    _rebase_pass.apply(pytket_circ)
+    AerBackend().rebase_pass().apply(pytket_circ)
 
     assert len(pytket_circ.free_symbols()) == 2
 
@@ -1003,7 +1005,7 @@ def _verify_single_q_rebase(
     u_before = backend.run_circuit(rotation_circ).get_unitary()
     circ = Circuit(1)
     circ.add_gate(OpType.TK1, [a, b, c], [0])
-    _rebase_pass.apply(circ)
+    backend.rebase_pass().apply(circ)
     u_after = backend.run_circuit(circ).get_unitary()
     return np.allclose(u_before, u_after)