+++ modules/pytket-qiskit/tests/qiskit_convert_test.py
@@ -1,4 +1,4 @@
-# Copyright 2019-2021 Cambridge Quantum Computing
+# Copyright 2019-2022 Cambridge Quantum Computing
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -653,3 +653,17 @@ def test_parameter_equality() -> None:
     final_circ.bind_parameters(param_dict)
 
     assert final_circ.parameters == circ.parameters
+
+
+# https://github.com/CQCL/pytket-extensions/issues/275
+def test_convert_multi_c_reg() -> None:
+    c = Circuit()
+    q0, q1 = c.add_q_register("q", 2)
+    c.add_c_register("c", 2)
+    [m0] = c.add_c_register("m", 1)
+    c.add_gate(OpType.X, [], [q1], condition_bits=[m0], condition_value=1)
+    c.CX(q0, q1)
+    c.add_gate(OpType.TK1, [0.5, 0.5, 0.5], [q0])
+    qcirc = tk_to_qiskit(c)
+    circ = qiskit_to_tk(qcirc)
+    assert circ.get_commands()[0].args == [m0, q1]