+++ modules/pytket-qsharp/pytket/extensions/qsharp/backends/azure_quantum.py
@@ -1,4 +1,4 @@
-# Copyright 2021 Cambridge Quantum Computing
+# Copyright 2021-2022 Cambridge Quantum Computing
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -257,7 +257,7 @@ class AzureBackend(_QsharpBaseBackend):
         try:
             return Backend.get_result(self, handle)
         except CircuitNotRunError:
-            timeout = kwargs.get("timeout")
+            timeout = cast(float, kwargs.get("timeout"))
             wait = kwargs.get("wait", 1.0)
             # Wait for job to finish; result will then be in the cache.
             end_time = (time.time() + timeout) if (timeout is not None) else None