+++ modules/pytket-qulacs/_metadata.py
@@ -1,2 +1,2 @@
-__extension_version__ = "0.15.0"
+__extension_version__ = "0.16.0"
 __extension_name__ = "pytket-qulacs"